+++ lib/one_logger.dart
@@ -4,7 +4,7 @@
 library simple_logger;
 
 export 'src/logger.dart';
-export 'src/helper.dart' hide dateFormat;
+export 'src/helper.dart' hide defaultDateFormat;
 export 'src/options/options.dart';
 export 'src/options/style.dart';
 export 'src/loki/options.dart';